[#552] Inactive groups filtering and deletion (#568)
* Implement the BB api

* Filter by days inactive

* Add a description field to the /gr/api/analytics/groups analytics API #567

* Send notification on delete group due to inactivity

* regenerate swag

* Small fix

* redo

* Additional fixes and doc rebuild

* Additional fixes

* Update api name
