+++ core/interfaces_bbs.go
@@ -1,0 +1,27 @@
+// Copyright 2022 Board of Trustees of the University of Illinois.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package core
+
+import (
+	"groups/core/model"
+)
+
+type bbsImpl struct {
+	app *Application
+}
+
+func (bbs *bbsImpl) OnUpdatedGroupExternalEntity(groupID string, operation model.ExternalOperation) error {
+	return bbs.app.onUpdatedGroupExternalEntity(groupID, operation)
+}
