+++ core/model/filters.go
@@ -36,6 +36,7 @@ type GroupsFilter struct {
 	Order            *string                        `json:"order"`  // order by category & name (asc desc)
 	Offset           *int64                         `json:"offset"` // result offset
 	Limit            *int64                         `json:"limit"`  // result limit
+	DaysInactive     *int64                         `json:"days_inactive"`
 } // @name GroupsFilter
 
 // PostsFilter Wraps all possible filters for getting group post call
