+++ docs/swagger.yaml
@@ -81,9 +81,15 @@ definitions:
         description: this is indicative and it's not required for update APIs
       date_created:
         type: string
+      date_events_updated:
+        type: string
       date_managed_membership_updated:
         type: string
       date_membership_updated:
+        type: string
+      date_polls_updated:
+        type: string
+      date_posts_updated:
         type: string
       date_updated:
         type: string
@@ -257,6 +263,8 @@ definitions:
       category:
         description: group category
         type: string
+      days_inactive:
+        type: integer
       exclude_my_groups:
         description: Exclude My groups
         type: boolean
@@ -943,6 +951,17 @@ definitions:
           $ref: '#/definitions/model.Post'
         type: array
     type: object
+  onGroupUpdatedRequestBody:
+    properties:
+      operation:
+        enum:
+        - event_update
+        - poll_update
+        - social_update
+        type: string
+    required:
+    - operation
+    type: object
   reportAbuseGroupPostRequestBody:
     properties:
       comment:
@@ -983,6 +1002,8 @@ definitions:
       date_created:
         type: string
       date_updated:
+        type: string
+      description:
         type: string
       hidden_for_search:
         type: boolean
@@ -2123,6 +2144,10 @@ paths:
         in: query
         name: limit
         type: string
+      - description: Deprecated - Filter by number of days inactive
+        in: query
+        name: days_inactive
+        type: string
       - description: Deprecated - instead use request body filter! include_hidden
           - Includes hidden groups if a search by title is performed. Possible value
           is true. Default false.
@@ -2182,6 +2207,10 @@ paths:
           result
         in: query
         name: limit
+        type: string
+      - description: Deprecated - Filter by number of days inactive
+        in: query
+        name: days_inactive
         type: string
       - description: Deprecated - instead use request body filter! include_hidden
           - Includes hidden groups if a search by title is performed. Possible value
@@ -2481,6 +2510,30 @@ paths:
                 $ref: '#/definitions/Group'
               type: array
             type: array
+      security:
+      - AppUserAuth: []
+      tags:
+      - BBS
+  /api/bbs/groups/{group_id}/date-updated:
+    put:
+      description: Recieves a callback notification from other BBS that group related
+        resource has been updated
+      operationId: OnGroupDateUpdated
+      parameters:
+      - description: group id
+        in: path
+        name: group-id
+        required: true
+        type: string
+      - description: 'Event type. Supported values: event_update, poll_update, social_update'
+        in: body
+        name: group-id
+        required: true
+        schema:
+          $ref: '#/definitions/onGroupUpdatedRequestBody'
+      responses:
+        "200":
+          description: OK
       security:
       - AppUserAuth: []
       tags:
