+++ driver/web/rest/admin_apis.go
@@ -901,6 +901,16 @@ func (h *AdminApisHandler) DeleteGroup(clientID string, current *model.User, w h
 		return
 	}
 
+	inactive := false
+	inactiveParam, ok := r.URL.Query()["inactive"]
+	if ok && len(inactiveParam[0]) > 0 {
+		val, err := strconv.ParseBool(inactiveParam[0])
+		if err != nil {
+			log.Printf("Invalid value for inactive parameter: %s", err.Error())
+		}
+		inactive = val
+	}
+
 	group, err := h.app.Services.GetGroupEntity(clientID, id)
 	if err != nil {
 		log.Println(err.Error())
@@ -913,7 +923,7 @@ func (h *AdminApisHandler) DeleteGroup(clientID string, current *model.User, w h
 		return
 	}
 
-	err = h.app.Services.DeleteGroup(clientID, current, id)
+	err = h.app.Admin.DeleteGroup(clientID, current, id, inactive)
 	if err != nil {
 		log.Printf("Error on deleting group - %s\n", err)
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
