+++ driver/web/rest/bbs_apis.go
@@ -5,12 +5,14 @@ import (
 	"errors"
 	"groups/core"
 	"groups/core/model"
+	"io"
 	"net/http"
 	"strings"
 
 	"github.com/gorilla/mux"
 	"github.com/rokwire/logging-library-go/v2/logs"
 	"github.com/rokwire/logging-library-go/v2/logutils"
+	"gopkg.in/go-playground/validator.v9"
 )
 
 // BBSApisHandler handles the rest BBS APIs implementation
@@ -162,4 +164,63 @@ func (h *BBSApisHandler) GetGroupsByGroupIDs(log *logs.Log, req *http.Request, u
 	}
 
 	return log.HTTPResponseSuccessJSON(data)
+}
+
+// onGroupUpdatedRequestBody response
+type onGroupUpdatedRequestBody struct {
+	Operation string `json:"operation" validate:"required,oneof=event_update poll_update social_update"`
+} // @name onGroupUpdatedRequestBody
+
+// OnGroupDateUpdated Recieves a callback notification from other BBS that group related resource has been updated
+// @Description Recieves a callback notification from other BBS that group related resource has been updated
+// @ID OnGroupDateUpdated
+// @Tags BBS
+// @Param group-id path string true "group id"
+// @Param group-id body onGroupUpdatedRequestBody true "Event type. Supported values: event_update, poll_update, social_update"
+// @Success 200
+// @Security AppUserAuth
+// @Router /api/bbs/groups/{group_id}/date-updated [put]
+func (h *BBSApisHandler) OnGroupDateUpdated(log *logs.Log, req *http.Request, user *model.User) logs.HTTPResponse {
+	params := mux.Vars(req)
+	groupID := params["group_id"]
+	if len(groupID) <= 0 {
+		return log.HTTPResponseErrorAction(logutils.ActionUpdate, logutils.TypePathParam, nil, errors.New("missing group_id"), http.StatusBadRequest, false)
+	}
+
+	data, err := io.ReadAll(req.Body)
+	if err != nil {
+		return log.HTTPResponseErrorAction(logutils.ActionUpdate, logutils.TypePathParam, nil, errors.New("unable to read request body"), http.StatusBadRequest, false)
+	}
+
+	var requestData onGroupUpdatedRequestBody
+	err = json.Unmarshal(data, &requestData)
+	if err != nil {
+		return log.HTTPResponseErrorAction(logutils.ActionUpdate, logutils.TypePathParam, nil, errors.New("unable to unmarshal request body"), http.StatusBadRequest, false)
+	}
+
+	//validate
+	validate := validator.New()
+	err = validate.Struct(requestData)
+	if err != nil {
+		return log.HTTPResponseErrorAction(logutils.ActionUpdate, logutils.TypePathParam, nil, errors.New("unable to validate request body"), http.StatusBadRequest, false)
+	}
+
+	var operationType model.ExternalOperation
+	switch requestData.Operation {
+	case "event_update":
+		operationType = model.ExternalOperationEventUpdate
+	case "poll_update":
+		operationType = model.ExternalOperationPollUpdate
+	case "social_update":
+		operationType = model.ExternalOperationSocialUpdate
+	default:
+		return log.HTTPResponseErrorAction(logutils.ActionUpdate, logutils.TypePathParam, nil, errors.New("unsupported operation. Expected: operation one of: event_update, poll_update, social_update"), http.StatusBadRequest, false)
+	}
+
+	err = h.app.BBS.OnUpdatedGroupExternalEntity(groupID, operationType)
+	if err != nil {
+		return log.HTTPResponseErrorAction(logutils.ActionGet, logutils.TypeError, nil, err, http.StatusBadRequest, false)
+	}
+
+	return log.HTTPResponseSuccess()
 }
